be/services/userProfileServices/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
gunicorn==21.2.0
psycopg[binary]==3.2.3
python-dotenv==1.0.0
supabase==2.3.0
gotrue==2.5.0
httpx==0.24.1
python-json-logger==2.0.7
redis
